Hook up rendy-descriptor

+++ wrench/src/main.rs
@@ -604,7 +604,7 @@ fn main() {
             adapter,
             surface,
             window_size: (dim.width, dim.height),
-            descriptor_count: args.value_of("descriptor_count").map(|d| d.parse::<usize>().unwrap()),
+            descriptor_count: args.value_of("descriptor_count").map(|d| d.parse::<u32>().unwrap()),
             cache_path,
             save_cache: true,
         }
